move sheet into member-details
