+++ src/components/members-table.tsx
@@ -74,25 +74,9 @@ import { Skeleton } from "./ui/skeleton";
 import InfiniteScroll from "react-infinite-scroll-component";
 import { RenewConfirm } from "./renew-confirm";
 import { toast } from "sonner";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-
-import {
-  Sheet,
-  SheetContent,
-  SheetHeader,
-  SheetTitle,
-  SheetTrigger,
-} from "@/components/ui/sheet";
-
 import { MemberDetails } from "./member-details";
 import { StatusBadge } from "./status-badge";
 import { SerializedMember } from "./pages/new-member";
-import { ScrollArea } from "@radix-ui/react-scroll-area";
 
 const columnHelper = createColumnHelper<Member>();
 const membersPerPage = 20;
@@ -152,20 +136,6 @@ export function DataTable({ search }: { search: string | null }) {
   const { t } = useTranslation();
   const [isRenewing, setIsRenewing] = useState<Record<string, undefined>>({});
   const queryClient = useQueryClient();
-
-  const [isMobile, setIsMobile] = useState(window.innerWidth <= 768);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setIsMobile(window.innerWidth <= 768);
-    };
-
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
 
   const renewMutation = useMutation({
     mutationFn: (variables: {
@@ -298,38 +268,15 @@ export function DataTable({ search }: { search: string | null }) {
 
           return (
             <div className="flex">
-              <Sheet>
-                <TooltipProvider>
-                  <Tooltip>
-                    <TooltipTrigger asChild>
-                      <SheetTrigger asChild>
-                        <Button size="icon" variant="ghost">
-                          <SquarePen className="w-5" />
-                        </Button>
-                      </SheetTrigger>
-                    </TooltipTrigger>
-                    <TooltipContent>
-                      {t("membersTable.editMember")}
-                    </TooltipContent>
-                  </Tooltip>
-                </TooltipProvider>
-                <SheetContent
-                  className={`overflow-y-scroll ${isMobile ? "w-full" : ""}`}
-                >
-                  <SheetHeader>
-                    <SheetTitle>
-                      <div className="flex gap-2 my-4">
-                        {`${row.original.name} ${row.original.surname}`}
-                        <StatusBadge status={row.original.status} />
-                      </div>
-                    </SheetTitle>
-                  </SheetHeader>
-                  <MemberDetails
-                    row={row.original}
-                    updateMutation={updateMutation}
-                  />
-                </SheetContent>
-              </Sheet>
+              <MemberDetails
+                row={row.original}
+                updateMutation={updateMutation}
+                tooltip={t("membersTable.editMember")}
+              >
+                <Button size="icon" variant="ghost">
+                  <SquarePen className="w-5" />
+                </Button>
+              </MemberDetails>
 
               <RenewConfirm
                 isOpenForbidden={isRenewForbidden}
@@ -343,25 +290,22 @@ export function DataTable({ search }: { search: string | null }) {
                 </Button>
               </RenewConfirm>
 
-              <TooltipProvider>
-                <Tooltip>
-                  <TooltipTrigger asChild>
-                    <Button size="icon" variant="ghost" disabled={!hasNote}>
-                      <MessageSquareText className="w-5" />
-                    </Button>
-                  </TooltipTrigger>
-                  <TooltipContent className="max-w-96">
-                    {t("memberDetails.noteLabel")}
-                    <p>{row.original.note}</p>
-                  </TooltipContent>
-                </Tooltip>
-              </TooltipProvider>
+              <MemberDetails
+                row={row.original}
+                updateMutation={updateMutation}
+                initialTab="note"
+                tooltip={row.original.note}
+              >
+                <Button size="icon" variant="ghost" disabled={!hasNote}>
+                  <MessageSquareText className="w-5" />
+                </Button>
+              </MemberDetails>
             </div>
           );
         },
       },
     ],
-    [t, isRenewing, isMobile, renewMutation, updateMutation],
+    [t, isRenewing, renewMutation, updateMutation],
   );
 
   const [columnVisibility, setColumnVisibility] = useState({
